+++ CHANGELOG.md
@@ -10,3 +10,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Queue module with functions for get queue, add and take message
 - Supervision tree
 - MessageServer for sending, receive result in Server
+- MnesiaDB support by Memento
+- uploading incompleted messages at the start